requirements.txt
asgiref==3.5.0
Django==3.2.12
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
screen==1.0.1
sqlparse==0.4.2
typing_extensions==4.1.1
xlwt==1.3.0
